Move mainFirst into the string_slice package
