+++ src/git/runner.go
@@ -724,7 +724,7 @@ func (r *Runner) LocalAndRemoteBranches() ([]string, error) {
 		i++
 	}
 	sort.Strings(result)
-	return MainFirst(result), nil
+	return stringslice.MainFirst(result), nil
 }
 
 // LocalBranches provides the names of all branches in the local repository,
@@ -748,7 +748,7 @@ func (r *Runner) LocalBranchesMainFirst() (result []string, err error) {
 	if err != nil {
 		return result, err
 	}
-	return MainFirst(sort.StringSlice(branches)), nil
+	return stringslice.MainFirst(sort.StringSlice(branches)), nil
 }
 
 // LocalBranchesWithDeletedTrackingBranches provides the names of all branches
